reinit entire project
Old configurations were noisy and just in the way for an unused template library and I'm switching out all the core libraries and doing major upgrades along the way so starting from a clean slate is a bit easier especially considering what this project is. This PR includes some work I did in a separate repository and copied over. Main things included:

- Configure new serverless services
- Install and configure esbuild and include bundle analyzer
- Get TypeScript working with some basic examples
- Add project READMEs
- Initialize monorepo structure with services directory
- Add global serverless.common.yml